requirements.txt
mcp[cli]
python-dotenv
google-api-python-client
google-auth
httpx